+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v1.0.0] - 2021-01-02
+### Added
+- Added support for hotwire/turbo
+
+### Changed
+Renamed project from alpine-turbolinks-adapter to alpine-turbo-adapter
 
 ## [v0.3.0] - 2020-09-27
 ### Added
@@ -42,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.0...HEAD
+[v1.0.0]: https://github.com/SimoTod/alpine-turbolinks-adapter/compare/v0.3.0...v1.0.0
 [v0.3.0]: https://github.com/SimoTod/alpine-turbolinks-adapter/compare/v0.2.1...v0.3.0
 [v0.2.1]: https://github.com/SimoTod/alpine-turbolinks-adapter/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/SimoTod/alpine-turbolinks-adapter/compare/v0.1.4...v0.2.0